Put the beta sample in the right file.
