+++ speech/cloud-client/beta_snippets_test.py
@@ -14,8 +14,10 @@
 import os
 
 from beta_snippets import (
-    transcribe_file_with_auto_punctuation, transcribe_file_with_enhanced_model,
-    transcribe_file_with_metadata)
+    transcribe_file_with_auto_punctuation,
+    transcribe_file_with_enhanced_model,
+    transcribe_file_with_metadata,
+    transcribe_file_with_diarization)
 
 RESOURCES = os.path.join(os.path.dirname(__file__), 'resources')
 
@@ -42,3 +44,11 @@ def test_transcribe_file_with_auto_punctuation(capsys):
     out, _ = capsys.readouterr()
 
     assert 'Okay. Sure.' in out
+
+
+def test_transcribe_diarization(capsys):
+    transcribe_file_with_diarization(
+        os.path.join(RESOURCES, 'Google_Gnome.wav'))
+    out, err = capsys.readouterr()
+
+    assert 'OK Google stream stranger things from Netflix to my TV' in out
